Revert "stats: split inline structs in actual types"
This reverts commit 71cd9e1dd69930e4584128f6646ecf35be2fefd4.
